+++ common/fmtc.spec
@@ -1,0 +1,97 @@
+################################################################################
+
+%define debug_package  %{nil}
+
+################################################################################
+
+Summary:        Simple utility for rendering fmtc formatted data
+Name:           fmtc
+Version:        0.1.0
+Release:        0%{?dist}
+Group:          Applications/System
+License:        Apache License, Version 2.0
+URL:            https://kaos.sh/fmtc
+
+Source0:        https://source.kaos.st/%{name}/%{name}-%{version}.tar.bz2
+
+BuildRoot:      %{_tmppath}/%{name}-%{version}-%{release}-root-%(%{__id_u} -n)
+
+BuildRequires:  golang >= 1.20
+
+Provides:       %{name} = %{version}-%{release}
+
+################################################################################
+
+%description
+Simple utility for rendering fmtc formatted data.
+
+################################################################################
+
+%prep
+%setup -q
+
+%build
+if [[ ! -d "%{name}/vendor" ]] ; then
+  echo "This package requires vendored dependencies"
+  exit 1
+fi
+
+pushd %{name}
+  go build %{name}.go
+  cp LICENSE ..
+popd
+
+%install
+rm -rf %{buildroot}
+
+install -dm 755 %{buildroot}%{_bindir}
+install -dm 755 %{buildroot}%{_mandir}/man1
+
+install -pm 755 %{name}/%{name} %{buildroot}%{_bindir}/
+
+./%{name}/%{name} --generate-man > %{buildroot}%{_mandir}/man1/%{name}.1
+
+%clean
+rm -rf %{buildroot}
+
+%post
+if [[ -d %{_sysconfdir}/bash_completion.d ]] ; then
+  %{name} --completion=bash 1> %{_sysconfdir}/bash_completion.d/%{name} 2>/dev/null
+fi
+
+if [[ -d %{_datarootdir}/fish/vendor_completions.d ]] ; then
+  %{name} --completion=fish 1> %{_datarootdir}/fish/vendor_completions.d/%{name}.fish 2>/dev/null
+fi
+
+if [[ -d %{_datadir}/zsh/site-functions ]] ; then
+  %{name} --completion=zsh 1> %{_datadir}/zsh/site-functions/_%{name} 2>/dev/null
+fi
+
+%postun
+if [[ $1 == 0 ]] ; then
+  if [[ -f %{_sysconfdir}/bash_completion.d/%{name} ]] ; then
+    rm -f %{_sysconfdir}/bash_completion.d/%{name} &>/dev/null || :
+  fi
+
+  if [[ -f %{_datarootdir}/fish/vendor_completions.d/%{name}.fish ]] ; then
+    rm -f %{_datarootdir}/fish/vendor_completions.d/%{name}.fish &>/dev/null || :
+  fi
+
+  if [[ -f %{_datadir}/zsh/site-functions/_%{name} ]] ; then
+    rm -f %{_datadir}/zsh/site-functions/_%{name} &>/dev/null || :
+  fi
+fi
+
+################################################################################
+
+%files
+%defattr(-,root,root,-)
+%doc LICENSE
+%{_mandir}/man1/%{name}.1.*
+%{_bindir}/%{name}
+
+################################################################################
+
+%changelog
+* Wed Nov 29 2023 Anton Novojilov <andy@essentialkaos.com> - 0.1.0-0
+- Initial build
